\docType{package}
\name{boutliers-package}
\alias{boutliers-package}
\title{The 'boutliers' package.}
\description{
Computational tools for implementing outlier detection and influence diagnostics for meta-analysis. Bootstrap distributions of the influence statistics are calculated, and the thresholds to determine outliers are provided.
}
\references{
Hedges, L. V., and Olkins, I. (1985). 
\emph{Statistical Methods for Meta-Analysis.} 
New York: Academic Press.

Noma, H., Gosho, M., Ishii, R., Oba, K., and Furukawa, T. A. (2020). 
Outlier detection and influence diagnostics in network meta-analysis. 
\emph{Research Synthesis Methods.}
\strong{11}(6): 891-902.
\doi{10.1002/jrsm.1455}

Viechtbauer, W., and Cheung, M. W. (2010).
Outlier and influence diagnostics for meta-analysis.
\emph{Research Synthesis Methods.}
\strong{1}(2): 112-125.
\doi{10.1002/jrsm.11}
}
