% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{finasteride}
\alias{finasteride}
\title{A multicenter clinical trial data assessing the treatment effect of finasteride for benign prostatic hyperplasia}
\format{A data frame with 29 rows and 7 variables}
\usage{
data(PPI)
}
\description{
\itemize{
  \item \code{center}: Center ID
  \item \code{n1}: Number of observations in finasteride group
  \item \code{m1}: Mean of the change of Boyarsky score from baseline in finasteride group
  \item \code{s1}: SD of the change of Boyarsky score from baseline in finasteride group
  \item \code{n0}: Number of observations in placebo group
  \item \code{m0}: Mean of the change of Boyarsky score from baseline in placebo group
  \item \code{s0}: SD of the change of Boyarsky score from baseline in placebo group
  }
}
\references{
Gormley, G. J., Stoner, E., Bruskewitz, R. C., et al. (1992). 
The effect of finasteride in men with benign prostatic hyperplasia. The Finasteride Study Group.
\emph{New England Journal of Medicine.}
\strong{327}: 1185-1191.
\url{https://www.nejm.org/doi/full/10.1056/NEJM199210223271701}

Gould, A. L. (1998). 
Multi-centre trial analysis revisited.
\emph{Statistics in Medicine.}
\strong{17}: 1779-1797.
}
\keyword{datasets}
