\docType{package}
\name{metaoutliers-package}
\alias{metaoutliers-package}
\alias{metaoutliers}
\title{The 'metaoutliers' package.}
\description{
A R package for implementing outlier detection and influence diagnostics for meta-analysis. Bootstrap distributions of the influence statistics are calculated, and the thresholds to determine outliers are clearly provided.
}
\references{
Hedges, L. V., and Olkins, I. (1985). 
\emph{Statistical Methods for Meta-Analysis.} 
New York: Academic Press.

Noma, H., Gosho, M., Ishii, R., Oba, K., and Furukawa, T. A. (2020). 
Outlier detection and influence diagnostics in network meta-analysis. 
arXiv: 1910.13080. 
\url{https://arxiv.org/abs/1910.13080}

Viechtbauer, W., and Cheung, M. W. (2010).
Outlier and influence diagnostics for meta-analysis.
\emph{Research Synthesis Methods.}
\strong{1}(2): 112-125.
\url{https://doi.org/10.1002/jrsm.11}
}
