% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{PPI}
\alias{PPI}
\title{Crocker et al. (2018)'s patient and public involvement (PPI) intervention data}
\format{A data frame with 21 rows and 5 variables}
\usage{
data(PPI)
}
\description{
\itemize{
  \item \code{ID}: Study ID
  \item \code{d1}: Number of events in PPI intervention group
  \item \code{n1}: Number of observations in PPI intervention group
  \item \code{d2}: Number of events in non-PPI intervention group
  \item \code{n2}: Number of observations in non-PPI intervention group
}
}
\references{
Crocker, J. C., Ricci-Cabello, I., Parker, A., Hirst, J. A., Chant, A., 
Petit-Zeman, S., Evans, D., Rees, S. (2018).
Impact of patient and public involvement on enrolment and 
retention in clinical trials: systematic review and meta-analysis.
\emph{BMJ.}
\strong{363}: k4738.
\url{https://doi.org/10.1136/bmj.k4738}
}
\keyword{datasets}
