\docType{package}
\name{boutliers-package}
\alias{boutliers-package}
\title{The 'boutliers' package.}
\description{
Computational tools for outlier detection and influence diagnostics in meta-analysis. Bootstrap distributions of influence statistics are computed, and explicit thresholds for identifying outliers are provided. These methods can also be applied to the analysis of influential centers or regions in multicenter or multiregional clinical trials.
}
\references{
Aoki, M., Noma, H., and Gosho, M. (2021). 
Methods for detecting outlying regions and influence diagnosis in multi-regional clinical trials. 
\emph{Biostatistics & Epidemiology.}
\strong{5}(1): 30-48.
\doi{10.1080/24709360.2021.1921944}

Hedges, L. V., and Olkins, I. (1985). 
\emph{Statistical Methods for Meta-Analysis.} 
New York: Academic Press.

Nakamura, R., and Noma, H. (2021). 
Detection of outlying centers and influence diagnostics for the analysis of multicenter clinical trials (in Japanese).
\emph{Japanese Journal of Biometrics.}
\strong{41}(2): 117-136.
\doi{10.5691/jjb.41.117}

Noma, H., Gosho, M., Ishii, R., Oba, K., and Furukawa, T. A. (2020). 
Outlier detection and influence diagnostics in network meta-analysis. 
\emph{Research Synthesis Methods.}
\strong{11}(6): 891-902.
\doi{10.1002/jrsm.1455}

Noma, H., Maruo, K., and Gosho, M. (2025). 
boutliers: R package of outlier detection and influence diagnostics for meta-analysis. 
\emph{medRxiv.}
\doi{10.1101/2025.09.18.25336125}

Viechtbauer, W., and Cheung, M. W. (2010).
Outlier and influence diagnostics for meta-analysis.
\emph{Research Synthesis Methods.}
\strong{1}(2): 112-125.
\doi{10.1002/jrsm.11}
}
